Fix code review findings: port_index truncation, EMA division, UTC time
Addresses findings from code review:

CRITICAL:
- port_index changed from u16 to u32 to handle SNMP ifIndex values up
  to 100000+ without silent truncation (Netgear uses 20000+)

MEDIUM:
- EMA baseline division: /(sample_count + 1) → /(sample_count + 1.0)
  to force float division in SQLite (was integer division)
- Baseline table added to wipe_device_port_data cleanup
- hour_of_week computation switched from Local to UTC to avoid DST
  bucket skipping/doubling

+++ crates/ion-drift-web/src/snmp_poller.rs
@@ -214,7 +214,7 @@ async fn poll_snmp_switch(
                 };
                 PortMetricEntry {
                     port_name: iface.canonical_name.clone(),
-                    port_index: iface.index as u16,
+                    port_index: iface.index,
                     rx_bytes: iface.rx_bytes,
                     tx_bytes: iface.tx_bytes,
                     rx_packets: iface.rx_packets,
